RavenFS refactoring - Moving methods from StorageOperationsTask to FileActions
