+++ Raven.Database/FileSystem/Util/SynchronizingFileStream.cs
@@ -1,12 +1,8 @@
-﻿using System.Collections.Specialized;
-using System.Security.Cryptography;
-using Raven.Abstractions.Util.Encryptors;
+﻿using Raven.Abstractions.Util.Encryptors;
 using Raven.Database.Extensions;
-using Raven.Database.FileSystem.Extensions;
-using Raven.Database.FileSystem.Infrastructure;
+using Raven.Database.FileSystem.Actions;
 using Raven.Database.FileSystem.Search;
 using Raven.Database.FileSystem.Storage;
-using Raven.Database.FileSystem.Storage.Esent;
 using Raven.Json.Linq;
 
 namespace Raven.Database.FileSystem.Util
@@ -17,7 +13,7 @@ namespace Raven.Database.FileSystem.Util
 
 		private SynchronizingFileStream(ITransactionalStorage transactionalStorage, string fileName,
 										StorageStreamAccess storageStreamAccess, RavenJObject metadata,
-										IndexStorage indexStorage, StorageOperationsTask operations)
+										IndexStorage indexStorage, FileActions operations)
 			: base(transactionalStorage, fileName, storageStreamAccess, metadata, indexStorage, operations)
 		{
 		    md5Hasher = Encryptor.Current.CreateHash();
@@ -48,7 +44,7 @@ namespace Raven.Database.FileSystem.Util
 		}
 
 		public static SynchronizingFileStream CreatingOrOpeningAndWriting(ITransactionalStorage storage, IndexStorage search,
-																		   StorageOperationsTask operationsTask,
+																		   FileActions operationsTask,
 																		   string fileName, RavenJObject metadata)
 		{
 			return new SynchronizingFileStream(storage, fileName, StorageStreamAccess.CreateAndWrite, metadata, search, operationsTask) { PreventUploadComplete = true };
